refactor: update/clarify types
