+++ src/dashboard_context/index.tsx
@@ -1,6 +1,18 @@
 import { createContext } from "react"
-import type { DashboardContextType } from "../types"
+import type { Widget } from "../widget"
 
-export const DashboardContext = createContext<DashboardContextType | undefined>(
+export interface DashboardContext {
+  widgets: Widget[]
+  addWidget: (widget: Widget) => void
+  updateWidget: (widget: Widget) => void
+  removeWidget: (id: string) => void
+  isModalOpen: boolean
+  selectedWidget: string | null
+  openModalForNewWidget: () => void
+  openModalForEdit: (widgetId: string) => void
+  closeModal: () => void
+}
+
+export const DashboardContext = createContext<DashboardContext | undefined>(
   undefined
 )
