+++ src/dashboard_context/use_dashboard.ts
@@ -1,6 +1,6 @@
 import { useContext } from "react"
 import { DashboardContext } from "../dashboard_context"
-import type { DashboardContextType } from "../types"
+import type { DashboardContext as DashboardContextType } from "../dashboard_context"
 
 export const useDashboard = (): DashboardContextType => {
   const context = useContext(DashboardContext)
