+++ src/quick_links/index.tsx
@@ -1,0 +1,14 @@
+export interface Link {
+  id: string
+  displayName: string
+  url: string
+  icon: {
+    name: string
+    type: "emoji" | "icon" | "image" | "logo"
+  }
+}
+
+export interface QuickLinks {
+  content: Link[]
+  setContent: (newContent: Link[]) => void
+}
